Add Factor and tests
