+++ src/EnhancedBayesianNetworks.jl
@@ -54,6 +54,7 @@ export children
 export discrete_ancestors
 export evaluate!
 export evaluate_with_envelopes
+export factorize
 export gplot
 export markov_blanket
 export markov_envelope
