+++ test/runtests.jl
@@ -18,7 +18,7 @@ include("networks/ebn/reduction/reduction.jl")
 include("networks/bn/bayesnet.jl")
 include("networks/cn/credalnet.jl")
 include("inference/inferencestate.jl")
-# include("inference/factors.jl")
+include("inference/factors.jl")
 # include("inference/factor_methods.jl")
 # include("inference/factor_algebra.jl")
 # include("inference/variableselimination.jl")
